Attempting to move templateInfo to caller
